+++ R/zzz.R
@@ -15,9 +15,9 @@
          #packageStartupMessage ("Attached: 'statuser' (Version: ",utils::packageVersion('statuser'),  ")") 
           
       #While developing:
-       # packageStartupMessage(
-        # "#######################################################\n",
-         #                     "This Version 2026 06 24 - 01.20PM\n" )
+        packageStartupMessage(
+         "#######################################################\n",
+                              "This Version 2026 07 08 - 01.20PM\n" )